requirements.txt
accelerate
google-generativeai
httpx
limits
openai
pandas==2.2.2
pydantic==2.9.2
pydeck
pytest==8.3.3
python-dotenv==1.0.1
slowapi
streamlit_js_eval
streamlit-extras==0.4.7
streamlit-tags==1.2.8
streamlit==1.40.0
torch --index-url https://download.pytorch.org/whl/cu118
torchaudio
torchvision
transformers